# app/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Wardrobe Fitter API",
    "description": "API documentation for the Wardrobe Fitter application"
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/user/{user}/collection": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user's collections",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User's collections"
          }
        }
      }
    },
    "/collection": {
      "get": {
        "tags": ["Collections"],
        "summary": "Get all collections",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of collections"
          }
        }
      }
    },
    "/collection/{collection}": {
      "get": {
        "tags": ["Collections"],
        "summary": "Get specific collection",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Collection details"
          }
        }
      },
      "delete": {
        "tags": ["Collections"],
        "summary": "Delete a collection",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Collection deleted successfully"
          }
        }
      }
    },
    "/collection/{collection}/favorite": {
      "post": {
        "tags": ["Collections"],
        "summary": "Favorite a collection",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Collection favorited successfully"
          }
        }
      }
    },
    "/collection/{collection}/unfavorite": {
      "delete": {
        "tags": ["Collections"],
        "summary": "Unfavorite a collection",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Collection unfavorited successfully"
          }
        }
      }
    },
    "/collection/{collection}/upvote": {
      "put": {
        "tags": ["Collections"],
        "summary": "Upvote a collection",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Collection upvoted successfully"
          }
        }
      }
    },
    "/collection/{collection}/unvote": {
      "delete": {
        "tags": ["Collections"],
        "summary": "Remove upvote from collection",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Upvote removed successfully"
          }
        }
      }
    },
...
    "/item": {
      "get": {
        "tags": ["Items"],
        "summary": "Get all items",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of items"
          }
        }
      }
    },
    "/item/{item}": {
      "get": {
        "tags": ["Items"],
        "summary": "Get specific item",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "item",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Item details"
          }
        }
      },
      "delete": {
        "tags": ["Items"],
        "summary": "Delete an item",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "item",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Item deleted successfully"
          }
        }
      }
    },
    "/item/{item}/favorite": {
      "post": {
        "tags": ["Items"],
        "summary": "Favorite an item",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "item",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Item favorited successfully"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}